"Implement shopping cart feature on products page"
This commit added several additional functionalities to the product page including a function to add a product to the shopping cart, which further required import of cartService, User model, and CartReqResp model. Each product now shows an add to cart button and clicking on it will add the said product to the user's shopping cart.

It also created and styled the Cart service, and updated the package.json to include lodash and updated several other packages. Additionally, the cart component has been temporarily commented out.